Add files via upload
The whole graphic is about a 3D-hairy ball and the background is a star field. The hairs and the ball can change the color with the mouse movement and so do the stars. The part of the star field is made with the reference of the coding challenge on Youtube.